Move game hooks to distinct folder
