+++ src/pages/Game/hooks/useAnimationLoop.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useRef } from 'react';
 
-import { Phase } from './types';
+import { Phase } from '../types';
 
 export const useAnimationLoop = (
   phase: Phase,
